# docs/02-design/specs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Design Copilot Backend
  version: 0.2.0
  description: |
    RD Design Copilot — AI-driven early-stage concept design decision platform.
    Integrates TRIZ divergence + KT decision convergence + evidence-driven 8-Gate process.

    ## Authentication
    All business endpoints require a valid JWT Bearer token issued by Supabase Auth.
    The sole exception is `POST /observability/web-vitals` (auth-less by design, sendBeacon cannot attach headers).

    ## Error Format
    All errors follow the standard FastAPI `HTTPValidationError` shape for 422, plus
    custom 400/404/502 responses documented per-endpoint.
  contact:
    name: RD Design Copilot Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.designcopilot.example.com
    description: Production (placeholder)
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/subsystems/suggest:
    post:
      operationId: subsystemsSuggest
      tags: [子系統 subsystems]
      summary: Suggest subsystems based on contradictions and brief
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubsystemSuggestRequest"
      responses:
        "200":
          description: Suggested subsystems with optional package map
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubsystemSuggestResponse"
        "502":
          description: Incomplete LLM response (fail-loud, no silent fallback)
          content:
            application/json:
              schema:
                type: object
                description: IncompleteLLMResponseError.to_dict() payload

  # v9: Migrated from /scamper/spatial-overlay
  /api/v1/subsystems/spatial-overlay:
    post:
      operationId: subsystemsSpatialOverlay
      tags: [子系統 subsystems]
      summary: Stateless what-if spatial overlay
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpatialOverlayRequest"
      responses:
        "200":
          description: Package map with overlay applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpatialOverlayResponse"

  # =========================================================================
  # 風險登錄 risks
  # =========================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth JWT token
  schemas:
...
    TrizSuggestion:
      type: object
      properties:
        path:
          type: string
          enum: [TC, PC, SuField, unknown]
          default: ""
        principle_number:
          type: integer
          nullable: true
        principle_name:
          type: string
          default: ""
        suggestion:
          type: string
        separation_principle:
          type: string
          default: ""
        affected_modules:
          type: array
          items:
            type: string
          default: []
        secondary_contradictions:
          type: array
          items:
            type: string
          default: []
      required: [suggestion]
...
    L1Surface:
      type: object
      properties:
        layer_role:
          type: string
          enum: [phenomenon, root_cause, structural_lens]
          default: phenomenon
        type:
          type: string
          enum: [TC]
          default: TC
        improving_param:
          type: integer
          nullable: true
        worsening_param:
          type: integer
          nullable: true
        candidate_principles:
          type: array
          items:
            type: integer
          default: []
        suggestions:
          type: array
          items:
            $ref: "#/components/schemas/TrizSuggestion"
          default: []
        depth_indicator:
          type: string
          enum: ["trade-off 改良", "根因突破", "功能鏈缺陷修補"]
          default: "trade-off 改良"
        evidence_level_floor:
          type: string
          enum: [E0, E1, E2, E3, E4]
          default: E1
        critic_trigger_l2:
          type: boolean
          default: false
        critic_reason:
          type: string
          default: ""
        critic_confidence:
          type: number
          default: 0.0
        status:
          type: string
          enum: [ran, skipped_quick_mode, skipped_condition, error]
          default: ran
    SeparationCandidate:
      type: object
      properties:
        type:
          type: string
          enum: [time, space, condition, whole_part]
        rationale:
          type: string
          default: ""
        confidence:
          type: number
          default: 0.0
      required: [type]
    DeepenLink:
      type: object
      properties:
        from_layer:
          type: string
          enum: [L1_surface]
          default: L1_surface
        from_tc_pair:
          type: array
          items:
            type: integer
            nullable: true
          minItems: 2
          maxItems: 2
        derived_physical_parameter:
          type: string
          default: ""
        contradiction_statement:
          type: string
          default: ""
        separation_type_candidates:
          type: array
          items:
            $ref: "#/components/schemas/SeparationCandidate"
          default: []
    L2RootCause:
      type: object
      properties:
        layer_role:
          type: string
          enum: [phenomenon, root_cause, structural_lens]
          default: root_cause
        type:
          type: string
          enum: [PC]
          default: PC
        triggered:
          type: boolean
          default: false
        trigger_reason:
          type: string
          default: ""
        deepen_link:
          $ref: "#/components/schemas/DeepenLink"
        suggestions:
          type: array
          items:
            $ref: "#/components/schemas/TrizSuggestion"
          default: []
        depth_indicator:
          type: string
          default: "根因突破"
        evidence_level_floor:
          type: string
          default: E1
        status:
          type: string
          enum: [ran, skipped_quick_mode, skipped_condition, error]
          default: skipped_condition
    SuFieldModel:
      type: object
      properties:
        S1:
          type: string
          default: ""
        S2:
          type: string
          default: ""
        F:
          type: string
          default: ""
        state:
          type: string
          enum: [incomplete, effective, harmful, insufficient, unknown]
          default: unknown
    L3StructuralCheck:
      type: object
      properties:
        layer_role:
          type: string
          default: structural_lens
        type:
          type: string
          enum: [SF]
          default: SF
        su_field_model:
          $ref: "#/components/schemas/SuFieldModel"
        matched_standard_solutions:
          type: array
          items:
            type: string
          default: []
        suggestions:
          type: array
          items:
            $ref: "#/components/schemas/TrizSuggestion"
          default: []
        supports_l1:
          type: string
          default: ""
        supports_l2:
          type: string
          default: ""
        standalone_value:
          type: string
          default: ""
        depth_indicator:
          type: string
          default: "功能鏈缺陷修補"
        evidence_level_floor:
          type: string
          default: E1
        status:
          type: string
          enum: [ran, skipped_quick_mode, skipped_condition, error]
          default: ran
    DifferentialPairAnalysis:
      type: object
      properties:
        on_solving_degree:
          type: string
          default: ""
        on_effort:
          type: string
          default: ""
        on_risk:
          type: string
          default: ""
        orthogonality:
          type: string
          default: ""
        synergy:
          type: string
          default: ""
    RecommendedRoute:
      type: object
      properties:
        primary:
          type: string
          default: ""
        fallback:
          type: string
          default: ""
        adopted_layers:
          type: array
          items:
            type: string
            enum: [L1, L2, L3]
          default: []
        rationale:
          type: string
          default: ""
    DifferentialAnalysis:
      type: object
      properties:
        l1_vs_l2:
          $ref: "#/components/schemas/DifferentialPairAnalysis"
        l1_vs_l3:
          $ref: "#/components/schemas/DifferentialPairAnalysis"
        l2_vs_l3:
          $ref: "#/components/schemas/DifferentialPairAnalysis"
        recommended_route:
          $ref: "#/components/schemas/RecommendedRoute"
    PhaseBDirective:
      type: object
      properties:
        same_contradiction_intra_layer_conflict:
          type: string
          enum: [skip, check]
          default: skip
        cross_contradiction_conflict:
          type: string
          enum: [skip, check]
          default: check
    LayeredTrizSolution:
      type: object
      properties:
        id:
          type: string
          example: LTS-EBIKE-012
        project_id:
          type: string
        contradiction_id:
          type: string
        contradiction_natural_description:
          type: string
          default: ""
        severity:
          type: string
          enum: [fatal, major, minor, unknown]
          default: unknown
        l1_surface:
          $ref: "#/components/schemas/L1Surface"
        l2_root_cause:
          $ref: "#/components/schemas/L2RootCause"
        l3_structural_check:
          $ref: "#/components/schemas/L3StructuralCheck"
        differential_analysis:
          $ref: "#/components/schemas/DifferentialAnalysis"
        phase_b_directive:
          $ref: "#/components/schemas/PhaseBDirective"
      required: [id, project_id, contradiction_id, l1_surface, l3_structural_check]
...
    InterfaceContract:
      type: object
      description: 6-dimensional interface contract between two coupled modules (camelCase wire format)
      properties:
        envelope:
          type: string
          default: ""
        loadPath:
          type: string
          default: ""
        thermalPath:
          type: string
          default: ""
        signalPath:
          type: string
          default: ""
        datumTolerance:
          type: string
          default: ""
        serviceability:
          type: string
          default: ""
        spatial:
          $ref: "#/components/schemas/SpatialEstimate"
    BBox:
      type: object
      properties:
        x_mm:
          type: number
        y_mm:
          type: number
        z_mm:
          type: number
        origin_mm:
          type: array
          items:
            type: number
          minItems: 3
          maxItems: 3
          default: [0.0, 0.0, 0.0]
        anchor:
          type: string
          default: ""
      required: [x_mm, y_mm, z_mm]
    SpatialEstimate:
      type: object
      properties:
        bbox:
          $ref: "#/components/schemas/BBox"
        mass_g:
          type: number
          nullable: true
        mounting_pattern:
          type: string
          default: ""
        reference_source:
          type: string
          default: ""
        confidence:
          type: string
          enum: [library, estimate, rd_confirmed]
          default: estimate
        rationale:
          type: string
          default: ""
...
    SuggestedSubsystem:
      type: object
      properties:
        name:
          type: string
        level:
          type: string
          enum: [system, module, component]
          default: module
        reason:
          type: string
          default: ""
        related_contradictions:
          type: array
          items:
            type: string
          default: []
        children:
          type: array
          items:
            $ref: "#/components/schemas/SuggestedSubsystem"
          default: []
        interface_contracts:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/InterfaceContract"
          default: {}
      required: [name]
    SubsystemSuggestRequest:
      type: object
      properties:
        project_id:
          type: string
        mission:
          type: string
        contradictions:
          type: array
          items:
            type: string
          default: []
        existing_subsystems:
          type: array
          items:
            type: string
          default: []
        layered_triz_solutions:
          type: array
          items:
            $ref: "#/components/schemas/LayeredTrizSolution"
          default: []
      required: [project_id, mission]
    PackageNode:
      type: object
      properties:
        name:
          type: string
        spatial:
          $ref: "#/components/schemas/SpatialEstimate"
        clashes:
          type: array
          items:
            type: string
          default: []
      required: [name, spatial]
    RequiredEnvelope:
      type: object
      properties:
        total_bbox_mm:
          type: array
          items:
            type: number
          minItems: 3
          maxItems: 3
        total_mass_g:
          type: number
        by_anchor:
          type: object
          additionalProperties:
            type: array
            items:
              type: number
            minItems: 3
            maxItems: 3
          default: {}
      required: [total_bbox_mm, total_mass_g]
    PackageMap:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/PackageNode"
          default: []
        required:
          $ref: "#/components/schemas/RequiredEnvelope"
        overlay_budget:
          type: object
          nullable: true
        overlay_violations:
          type: array
          items:
            type: string
          default: []
        svg:
          type: string
          default: ""
        table_md:
          type: string
          default: ""
        notes:
          type: array
          items:
            type: string
          default: []
    SubsystemSuggestResponse:
      type: object
      properties:
        subsystems:
          type: array
          items:
            $ref: "#/components/schemas/SuggestedSubsystem"
        package_map:
          $ref: "#/components/schemas/PackageMap"
      required: [subsystems]
    SpatialOverlayRequest:
      type: object
      properties:
        project_id:
          type: string
        subsystems:
          type: array
          items:
            $ref: "#/components/schemas/SuggestedSubsystem"
          default: []
        overlay:
          type: object
          default: {}
      required: [project_id]
    SpatialOverlayResponse:
      type: object
      properties:
        package_map:
          $ref: "#/components/schemas/PackageMap"
      required: [package_map]
...